修改文档权限批量修改中的管理员唯一判断

+++ app/Model/Logic/DocumentPermissionLogic.php
@@ -36,14 +36,6 @@ class DocumentPermissionLogic extends BaseLogic
 		$documentPermission->document_id = $documentId;
 		$documentPermission->permission = $permission;
 
-		if ($documentPermission->permission == DocumentPermission::MANAGER_PERMISSION) {
-			//只能添加一个
-			$exist = DocumentPermission::query()->where('document_id', '=', $documentId)->where('permission', '=', $permission)->first();
-			if ($exist) {
-				throw new \RuntimeException('该文档的管理员已存在');
-			}
-		}
-
 		if (!$documentPermission->save()) {
 			throw new \RuntimeException('文档权限添加失败');
 		}
@@ -57,17 +49,11 @@ class DocumentPermissionLogic extends BaseLogic
 			throw new \RuntimeException('该文档不存在');
 		}
 
-		//如果当前权限是管理员，删除已存在的管理员
-		if ($permission == DocumentPermission::MANAGER_PERMISSION) {
-			$documentManager = DocumentPermission::query()->where('document_id', '=', $documentId)->where('permission', '', $permission)->first();
-			if ($documentManager) {
-				$documentManager->delete();
-			}
-		}
-
+		//如果permission为空，删除对应权限
 		$documentPermission = $this->getByDocIdAndUid($documentId, $userId);
-		if ($documentPermission && !$permission) {
-			$documentPermission->delete();
+		if (!$permission) {
+			$documentPermission && $documentPermission->delete();
+			return true;
 		}
 		if (!$documentPermission) {
 			$documentPermission = new DocumentPermission();
